alvera/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.3
groq==0.11.0
python-dotenv==1.0.1
iyzipay==1.0.4
requests==2.32.3
